agents/Simple Data Analysis Agent (PydanticAI)/requirements.txt
pydantic-ai>=0.0.20
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
streamlit>=1.28.0

